@@ -2,7 +2,7 @@
     "require": {
         "php-cs-fixer/shim": "^3.0",
         "friendsoftwig/twigcs": "^6.0",
-        "symfony/phpunit-bridge": "^6.4|^7.0|^8.0",
+        "symfony/phpunit-bridge": "^7.4|^8.0",
         "phpstan/phpstan": "^2.0",
         "phpstan/extension-installer": "^1.4",
         "phpstan/phpstan-symfony": "^2.0"
