@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Siganushka\GenericBundle\Tests\Utils;
 
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 use Siganushka\GenericBundle\Utils\ExcelUtils;
 
 class ExcelUtilsTest extends TestCase
 {
-    /**
-     * @dataProvider columnDimensionProvider
-     */
+    #[DataProvider('columnDimensionProvider')]
     public function testGenerateColumnDimension(int $number, string $columnDimension): void
     {
         static::assertSame($columnDimension, ExcelUtils::generateColumnDimension($number));
