@@ -12,8 +12,8 @@ $finder = (new PhpCsFixer\Finder())
 return (new PhpCsFixer\Config())
     ->setParallelConfig(PhpCsFixer\Runner\Parallel\ParallelConfigFactory::detect())
     ->setRules([
-        '@PHP8x1Migration' => true,
-        '@PHP8x1Migration:risky' => true,
+        '@PHP8x2Migration' => true,
+        '@PHP8x2Migration:risky' => true,
         '@PHPUnit9x1Migration:risky' => true,
         '@Symfony' => true,
         '@Symfony:risky' => true,
